Reorganize and rename, and start UserAuth stuff.
