+++ kitsune/sumo/static/sumo/js/utils/apiFetch.es6.js
@@ -1,18 +1,21 @@
 /* globals k:false */
-import 'fetch'; // polyfill
+
+// import 'fetch'; // polyfill
+import cx from 'classnames';
 
 export default function apiFetch(url, options) {
+  if (url.indexOf('?') === -1) {
+    url += '?format=json&';
+  } else {
+    url += '&format=json&';
+  }
+
   if ('data' in options) {
     if ('body' in options) {
       throw new Error('Only pass one of `options.data` and `options.body`.');
     }
     let method = (options.method || 'get').toLowerCase();
     if (method === 'get' || method === 'head') {
-      if (url.indexOf('?') === -1) {
-        url += '?';
-      } else {
-        url += '&';
-      }
       /* The slice is to remove the ? that that
        * `queryParamStringFromDict` includes, since it was added above. */
       url += k.queryParamStringFromDict(options.data).slice(1);
