+++ kitsune/sumo/static/sumo/js/utils/index.js
@@ -1,0 +1,9 @@
+export apiFetch from './apiFetch.es6';
+export constantMap from './constantMap.es6';
+export scrollTo from './scrollTo.es6';
+
+export default {
+  apiFetch,
+  constantMap,
+  scrollTo,
+};
