+++ kitsune/sumo/static/sumo/less/main.less
@@ -406,7 +406,8 @@ input[type=button],
 }
 
 .btn {
-  &:disabled {
+  &:disabled,
+  &.disabled {
     opacity: 0.5;
   }
 }
