+++ kitsune/users/static/users/js/constants/UserAuthConstants.es6.js
@@ -1,0 +1,6 @@
+import constantMap from '../../../sumo/js/utils/constantMap.es6';
+
+export const actionTypes = constantMap([
+  'AUTH_LOGGED_IN',
+  'AUTH_LOGGED_OUT',
+]);
